test-kw/requirements.txt
pymongo==4.6.0
boto3==1.34.0
python-dotenv==1.0.0